Address PR feedback + remove unnecessary casts/using directives.

+++ src/System.Memory/src/System/Buffers/Text/Utf8Formatter/Utf8Formatter.Decimal.F.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
 
 namespace System.Buffers.Text
 {
@@ -69,7 +68,7 @@ namespace System.Buffers.Text
                 int numDigitsEmitted = 0;
                 if (scale < 0)
                 {
-                    int numLeadingZeroesToEmit = Math.Min((int)precision, -scale);
+                    int numLeadingZeroesToEmit = Math.Min(precision, -scale);
                     for (int i = 0; i < numLeadingZeroesToEmit; i++)
                     {
                         buffer[dstIndex++] = (byte)'0';
